\name{matrixSample}
\alias{matrixSample}
\docType{data}
\title{
matrixSample - a matrix
}
\description{
A matrix representation of the dataframe sample.}
\usage{data(matrixSample)}
\format{
  a matrix with 87 lines and 80 columns.
}
\examples{
data(matrixSample)
STdiag(matrixSample,log=TRUE)
}
\keyword{datasets}
