\name{STdiag-package}
\alias{STdiag-package}
\docType{package}
\title{
Functions to Plot a Structure-Time Diagram}
\description{
The STdiag package is designed to provide efficient tools to simply plot a structure-time diagram to represent dynamics of a structured time series.}
\details{
\tabular{ll}{
Package: \tab STdiag\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-23\cr
License: \tab GPL (>= 2)\cr
}
Use STdiag function to draw a structure-time diagram and Interpolation to smooth data in both x and y directions.
}
\author{
Vincent Le Bourlot

Maintainer: Vincent Le Bourlot <vincent.le.bourlot@ens.fr>
}
\keyword{ package }
