\name{StructureTimeDiagram-package}
\alias{StructureTimeDiagram-package}
\alias{StructureTimeDiagram}
\docType{package}
\title{
Plot a structure time diagram}
\description{
This package provides a function to plot a structure-time diagram and an bi-directional interpolation function. }
\details{
\tabular{ll}{
Package: \tab StructureTimeDiagram\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-23\cr
License: \tab GPL (>= 2)\cr
}
Use STdiag function to draw a structure-time diagram and Interpolation to smooth data in both x and y directions.
}
\author{
Vincent Le Bourlot

Maintainer: Vincent Le Bourlot <vincent.le.bourlot@ens.fr>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
