\name{STdiag-package}
\alias{STdiag-package}
\docType{package}
\title{
A package to plot structured population dynamics}
\description{
This package provides functions to plot a structure-time diagram and smooth or interpolate the data. }
\details{
\tabular{ll}{
Package: \tab STdiag\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-23\cr
License: \tab GPL (>= 2)\cr
}
Use STdiag function to draw a structure-time diagram and Interpolation to smooth data in both x and y directions.
}
\author{
Vincent Le Bourlot

Maintainer: Vincent Le Bourlot <vincent.le.bourlot@ens.fr>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
